separated MOI
