+++ Project.toml
@@ -15,6 +15,7 @@ KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 oneAPI = "8f75cd03-7ff8-4ecb-9b8f-daf728133b1b"
+MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 
 [extensions]
 ExaModelsAMDGPU = "AMDGPU"
@@ -23,6 +24,7 @@ ExaModelsKernelAbstractions = "KernelAbstractions"
 ExaModelsOneAPI = "oneAPI"
 ExaModelsSpecialFunctions = "SpecialFunctions"
 ExaModelsJuMP = "JuMP"
+ExaModelsMOI = "MathOptInterface"
 
 [compat]
 AMDGPU = "0.7"
@@ -33,6 +35,7 @@ SolverCore = "0.3"
 SpecialFunctions = "2"
 julia = "1.9"
 oneAPI = "1"
+MathOptInterface = "1.19"
 
 [extras]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
